Fixed up the UI
