+++ app/src/main/java/com/jakester/twitterapp/adapter/TwitterFragmentPagerAdapter.java
@@ -23,7 +23,7 @@ public class TwitterFragmentPagerAdapter extends FragmentPagerAdapter {
 
     @Override
     public int getCount() {
-        return 2;
+        return tabTitles.length;
     }
 
     @Override
