+++ app/src/main/java/com/jakester/twitterapp/fragments/HomeTimelineFragment.java
@@ -2,10 +2,12 @@ package com.jakester.twitterapp.fragments;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
+import android.support.design.widget.FloatingActionButton;
 import android.support.v4.app.Fragment;
 import android.support.v4.widget.SwipeRefreshLayout;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -13,18 +15,26 @@ import android.view.ViewGroup;
 import com.jakester.twitterapp.R;
 import com.jakester.twitterapp.activities.HomeTimelineActivity;
 import com.jakester.twitterapp.adapter.TweetAdapter;
+import com.jakester.twitterapp.application.TwitterApplication;
 import com.jakester.twitterapp.listener.EndlessScrollListener;
 import com.jakester.twitterapp.managers.InternetManager;
 import com.jakester.twitterapp.models.Tweet;
 import com.jakester.twitterapp.network.TwitterClient;
+import com.loopj.android.http.JsonHttpResponseHandler;
+import com.raizlabs.android.dbflow.sql.language.SQLite;
+
+import org.json.JSONArray;
+import org.json.JSONObject;
 
 import java.util.ArrayList;
+
+import cz.msebera.android.httpclient.Header;
 
 /**
  * Created by Jake on 10/4/2017.
  */
 
-public class HomeTimelineFragment extends Fragment {
+public class HomeTimelineFragment extends Fragment  implements NewTweetDialogFragment.FilterDialogListener {
 
     ArrayList<Tweet> tweets;
     TweetAdapter mAdapter;
@@ -48,7 +58,7 @@ public class HomeTimelineFragment extends Fragment {
                 // Triggered only when new data needs to be appended to the list
                 // Add whatever code is needed to append new items to the bottom of the list
                 if(InternetManager.getInstance(getContext()).isInternetAvailable()) {
-                    ((HomeTimelineActivity)getActivity()).populateTimeline(page+1,false);
+                    populateTimeline(page+1,false);
                 }
                 else{
                     ((HomeTimelineActivity)getActivity()).showNoInternetDialog();
@@ -65,7 +75,7 @@ public class HomeTimelineFragment extends Fragment {
                 // Your code to refresh the list here.
                 // Make sure you call swipeContainer.setRefreshing(false)
                 // once the network request has completed successfully.
-                ((HomeTimelineActivity)getActivity()).populateTimeline(0, true);
+                populateTimeline(0, true);
             }
         });
         // Configure the refreshing colors
@@ -77,12 +87,107 @@ public class HomeTimelineFragment extends Fragment {
         mAdapter = new TweetAdapter(getContext());
         mTweetRecycler.setAdapter(mAdapter);
 
+
+
+        FloatingActionButton fab = (FloatingActionButton) v.findViewById(R.id.fab);
+        fab.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                ((HomeTimelineActivity)getActivity()).showNewTweetDialog();
+            }
+        });
+
         return v;
+    }
+
+    public void populateTimeline(int page, final boolean refresh){
+        client.getHomeTimeline(page, new JsonHttpResponseHandler() {
+            @Override
+            public void onSuccess(int statusCode, Header[] headers, JSONObject response) {
+                Log.d("OBJECT", response.toString());
+            }
+
+            @Override
+            public void onSuccess(int statusCode, Header[] headers, JSONArray response) {
+                addItems(Tweet.fromJson(response),refresh);
+
+            }
+
+            @Override
+            public void onFailure(int statusCode, Header[] headers, Throwable throwable, JSONObject errorResponse) {
+                Log.d("TwitterClient", errorResponse.toString());
+                throwable.printStackTrace();
+            }
+
+            @Override
+            public void onFailure(int statusCode, Header[] headers, Throwable throwable, JSONArray errorResponse) {
+                Log.d("TwitterClient", errorResponse.toString());
+                throwable.printStackTrace();
+            }
+
+            @Override
+            public void onFailure(int statusCode, Header[] headers, String responseString, Throwable throwable) {
+                Log.d("TwitterClient", responseString);
+                throwable.printStackTrace();
+            }
+        });
+    }
+
+    @Override
+    public void onFinishFilterDialog(final Tweet tweet) {
+        client.postTweet(tweet.getTweet(), new JsonHttpResponseHandler(){
+            @Override
+            public void onSuccess(int statusCode, Header[] headers, JSONObject response) {
+                Log.d("OBJECT", response.toString());
+                addTweet(tweet);
+            }
+
+            @Override
+            public void onSuccess(int statusCode, Header[] headers, JSONArray response) {
+                Log.d("ARRAY", response.toString());
+                addTweet(tweet);
+            }
+
+            @Override
+            public void onFailure(int statusCode, Header[] headers, Throwable throwable, JSONObject errorResponse) {
+                Log.d("TwitterClient", errorResponse.toString());
+                throwable.printStackTrace();
+            }
+
+            @Override
+            public void onFailure(int statusCode, Header[] headers, Throwable throwable, JSONArray errorResponse) {
+                Log.d("TwitterClient", errorResponse.toString());
+                throwable.printStackTrace();
+            }
+
+            @Override
+            public void onFailure(int statusCode, Header[] headers, String responseString, Throwable throwable) {
+                Log.d("TwitterClient", responseString);
+                throwable.printStackTrace();
+            }
+        });
+
+    }
+
+    @Override
+    public void onActivityCreated(@Nullable Bundle savedInstanceState) {
+        super.onActivityCreated(savedInstanceState);
+        client = TwitterApplication.getRestClient();
     }
 
     @Override
     public void onResume() {
         super.onResume();
+        client = TwitterApplication.getRestClient();
+        if(client.isAuthenticated() && InternetManager.getInstance(getActivity()).isInternetAvailable()) {
+            populateTimeline(0, false);
+        }
+        if(!InternetManager.getInstance(getActivity()).isInternetAvailable()){
+            ArrayList<Tweet> tweets = (ArrayList<Tweet>) SQLite.select()
+                    .from(Tweet.class).queryList();
+            mAdapter.addTweets(tweets);
+            swipeContainer.setRefreshing(false);
+        }
     }
