+++ app/src/main/res/layout/activity_home_timeline.xml
@@ -6,9 +6,21 @@
     android:layout_height="match_parent"
     tools:context="com.jakester.twitterapp.activities.HomeTimelineActivity">
 
-    <android.support.design.widget.AppBarLayout
+    <android.support.v4.view.ViewPager
+        android:id="@+id/viewpager"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
+        android:layout_alignParentBottom="true"
+        android:adjustViewBounds="true"
+        app:layout_behavior="@string/appbar_scrolling_view_behavior"
+        android:background="@android:color/white" />
+
+
+    <android.support.design.widget.AppBarLayout
+        android:id="@+id/ab_layout"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_above="@id/viewpager"
         android:theme="@style/AppTheme.AppBarOverlay">
 
         <android.support.v7.widget.Toolbar
@@ -38,17 +50,22 @@
         </android.support.v7.widget.Toolbar>
 
 
+        <android.support.design.widget.TabLayout
+            android:id="@+id/sliding_tabs"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            app:layout_scrollFlags="snap"
+            app:tabGravity="center"
+            app:tabTextColor="@color/primaryTextColor"
+            app:tabSelectedTextColor="@color/primaryTextColor"
+            app:tabMode="fixed" />
+
+
     </android.support.design.widget.AppBarLayout>
 
 
-    <include layout="@layout/content_main_twitter" />
 
-    <android.support.design.widget.FloatingActionButton
-        android:id="@+id/fab"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom|end"
-        android:layout_margin="@dimen/fab_margin"
-        app:srcCompat="@drawable/ic_compose_tweet" />
+
+
 
 </android.support.design.widget.CoordinatorLayout>
