+++ app/src/main/res/layout/content_main_twitter.xml
@@ -8,7 +8,8 @@
     <android.support.design.widget.TabLayout
         android:id="@+id/sliding_tabs"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
+        android:layout_height="?attr/actionBarSize"
+        android:layout_gravity="bottom"
         app:tabGravity="center"
         app:tabMode="fixed" />
